Fully rewrote the user list page
- Fixed variable in pagination component
- Rewrote user list page to have something working
